requirements.txt
pykernel==6.25.1
ipython==8.14.0
ipython_genutils==0.2.0
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupytext==1.15.0
Markdown==3.4.4
markdown-it-py==3.0.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
mdit-py-plugins==0.4.0
mpmath==1.3.0
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python-headless==4.8.0.76
pandas==2.0.3
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2023.7.23
pyparsing==3.1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.49
python-json-logger==2.0.7
pytz==2023.3
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.1
scikit-learn==1.3.0
scipy==1.11.1
torch==2.0.1
torcheval==0.0.6
torchtnt==0.2.0
torchvision==0.15.2